requirements.txt.txt
numpy==1.21.5
pandas==1.4.2
Pillow==9.0.1
streamlit==1.10.0
sectionproperties==2.1.4
matplotlib==3.4.2